# swagger.yaml
openapi: 3.0.0
info:
  title: Damnplay API Documentation
  version: 1.0.0
  description: API documentation for Damnplay's contest, game, leaderboard, user, and wallet modules.
components:
  securitySchemes:
    AccessTokenAuth:
      type: apiKey
      in: header
      name: access-token  # Define the name of the custom header

    HeaderAuth:
      type: apiKey
      in: header
      name: Authorization  # Define the name of the custom header  
    
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
servers:
  - url: http://localhost:5000
    description: Local development server
paths:
  /damnplay/contest/create:
    post:
      summary: Create a new contest
      description: |
        This endpoint allows creating a new contest. 
        - Requires a valid `access-token`.
        - User must have an admin role.
      security:
        - bearerAuth: []
        - AccessTokenAuth: []     # Enforces the `access-token` header
      requestBody:
        description: Contest creation details
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                entry_fee:
                  type: number
                  description: Entry fee for the contest
      responses:
        '200':
          description: Contest created successfully
        '400':
          description: Entry fee is required
        '500':
          description: Server error
  /damnplay/contest/join:
    post:
      summary: Join a contest
      description: Deducts entry fee from the user's wallet.
      security:
        - bearerAuth: []
        - AccessTokenAuth: []
      requestBody:
        description: Contest joining details
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                contest_id:
                  type: string
                  description: ID of the contest to join
      responses:
        '200':
          description: Joined contest successfully
        '400':
          description: Contest ID is required
        '404':
          description: Contest not found
        '500':
          description: Server error
  /damnplay/contest/active:
    get:
      summary: Retrieve active contests
      description: No authentication required.
      responses:
        '200':
          description: Active contests retrieved successfully
        '500':
          description: Server error
  /damnplay/contest/cancel:
    post:
      summary: Cancel a contest
      description: |
        This endpoint Cancel Contest and refund. 
        - Requires a valid `access-token`.
        - User must have an admin role.
      security:
        - bearerAuth: []
        - AccessTokenAuth: []    # Enforces the `access-token` header
      requestBody:
        description: Contest cancellation details
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                contest_id:
                  type: string
                  description: ID of the contest to cancel
      responses:
        '200':
          description: Contest canceled successfully
        '400':
          description: Contest ID is required
        '500':
          description: Server error
  /damnplay/game/games:
    get:
      summary: List games with filters and pagination
      parameters:
        - in: query
          name: category
          schema:
            type: string
          description: Filter by game category
        - in: query
          name: min_popularity
          schema:
            type: number
            default: 0
          description: Minimum popularity score
        - in: query
          name: max_popularity
          schema:
            type: number
            default: 100
          description: Maximum popularity score
        - in: query
          name: min_rating
          schema:
            type: number
            default: 0
          description: Minimum rating
        - in: query
          name: max_rating
          schema:
            type: number
            default: 5
          description: Maximum rating
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: Page number for pagination
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
          description: Number of items per page
      responses:
        '200':
          description: List of games retrieved successfully
        '400':
          description: Invalid query parameters
        '500':
          description: Server error

    post:
      summary: Create a new game
      description: Add metadata for a new game.
      requestBody:
        description: Game details
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: Title of the game
                category:
                  type: string
                  description: Category of the game
                description:
                  type: string
                  description: Description of the game
                release_year:
                  type: integer
                  description: Release year of the game
                popularity:
                  type: number
                  description: Popularity score of the game
                average_rating:
                  type: number
                  description: Average rating of the game
      responses:
        '201':
          description: Game created successfully
        '400':
          description: Missing or invalid fields
        '500':
          description: Server error
...